// client/src/app/shared/models/interfaces/triangle-string.interface.ts
export interface TriangleStringInterface {
  side_a: string;
  side_b: string;
  side_c: string;
  angle_a: string;
  angle_b: string;
  angle_c: string;
  bisector_a: string;
  bisector_b: string;
  bisector_c: string;
  altitude_a: string;
  altitude_b: string;
  altitude_c: string;
  median_a: string;
  median_b: string;
  median_c: string;
  radius_R: string;
  radius_r: string;
  radius_ra: string;
  radius_rb: string;
  radius_rc: string;
}